Temporarily removed security classes, removed old OCR and Response DTOs, created new classes to properly handle requests with Python microservice for PDF paystub data retrieval

+++ ezstub-backend/src/main/java/ezstub_backend/controller/AnalyticsController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @RestController
@@ -21,13 +22,13 @@ public class AnalyticsController {
 
     // Get total spending per month
     @GetMapping("/spending/user/{userId}/month/{month}")
-    public Map<String, Double> getTotalSpending(@PathVariable Long userId, @PathVariable String month) {
+    public Map<String, BigDecimal> getTotalSpending(@PathVariable Long userId, @PathVariable String month) {
         return analyticsService.getTotalSpending(userId, month);
     }
 
     // Get spending by category
     @GetMapping("/categories/user/{userId}/month/{month}")
-    public Map<String, Double> getSpendingByCategory(@PathVariable Long userId, @PathVariable String month) {
+    public Map<String, BigDecimal> getSpendingByCategory(@PathVariable Long userId, @PathVariable String month) {
         return analyticsService.getSpendingByCategory(userId, month);
     }
