+++ ezstub-backend/src/main/java/ezstub_backend/dto/BudgetDTO.java
@@ -3,6 +3,8 @@ package ezstub_backend.dto;
 import ezstub_backend.model.enums.ExpenseCategory;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -12,7 +14,7 @@ public class BudgetDTO {
 
     private Long id;
     private ExpenseCategory category;
-    private Double limitAmount;
+    private BigDecimal limitAmount;
     private String month;
     private Long userId;
 }
