+++ ezstub-backend/src/main/java/ezstub_backend/service/AnalyticsService.java
@@ -1,12 +1,13 @@
 package ezstub_backend.service;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface AnalyticsService {
 
-    Map<String, Double> getTotalSpending(Long userId, String month);
+    Map<String, BigDecimal> getTotalSpending(Long userId, String month);
 
-    Map<String, Double> getSpendingByCategory(Long userId, String month);
+    Map<String, BigDecimal> getSpendingByCategory(Long userId, String month);
 
     Map<String, Object> getBudgetSummary(Long userId, String month);
 }
